work on some user specifications

+++ base/views.py
@@ -1,33 +1,41 @@
+from django.http import HttpResponse
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth.models import User
 from .models import Room, Topic
 from django.contrib.auth import authenticate, login, logout
-
-from django.db.models import Q
 from .forms import RoomForm
 # Create your views here.
 
 
-def login_view(request): 
-    if request.method == 'POST':
-        username = request.POST.get('username')
-        password = request.POST.get('password')
+def login_view(request):  
+  
+  if request.user.is_authenticated:
+    return redirect('home')
+   
+  if request.method == 'POST':
+      username = request.POST.get('username')
+      password = request.POST.get('password')
 
-        try:
-            user = User.objects.get(username=username)
-        except User.DoesNotExist:
-            messages.error(request, "User does not exist.")
-            return redirect('login')  # redirect back to the login page
-        user = authenticate(request, username=username, password=password)
-        if user is not None:
-          login(request, user)
-          return redirect('home')
-        else: 
-          messages.error(request, 'User OR Password does not exist')
-    return render(request, 'base/login_form.html')
+      try:
+          user = User.objects.get(username=username)
+      except User.DoesNotExist:
+          messages.error(request, "User does not exist.")
+          return redirect('login')  # redirect back to the login page
+      user = authenticate(request, username=username, password=password)
+      if user is not None:
+        login(request, user)
+        return redirect('home')
+      else: 
+        messages.error(request, 'User OR Password does not exist')
+  return render(request, 'base/login_form.html')
 
 
+def logout_view(request):
+  logout(request)
+  return redirect('login')
 
 def home(request): 
   q = request.GET.get('q') or ''
@@ -49,7 +57,7 @@ def room(request, pk):
   return render(request, 'base/room.html', context)
 
 
-
+@login_required(login_url='/login')
 def create_room(request): 
   form = RoomForm()
   if request.method == 'POST':
@@ -61,10 +69,13 @@ def create_room(request):
   return render(request, 'base/room-form.html', context)
 
 
-
+@login_required(login_url='/login')
 def update_room(request, pk): 
   room = Room.objects.get(id = pk)
   form = RoomForm(instance=room)
+  
+  if request.user != room.host:
+    return HttpResponse('You can only edit your own posts')
   
   if request.method == 'POST':
     form = RoomForm(request.POST, instance=room)
@@ -75,9 +86,13 @@ def update_room(request, pk):
   return render(request, 'base/room-form.html', context)
 
 
-
+@login_required(login_url='/login')
 def delete_room(request, pk):
   room = Room.objects.get(id = pk)
+  
+  if request.user != room.host:
+    return HttpResponse('You can only delete your own posts')
+  
   if request.method == 'POST':
     room.delete()
     return redirect('home')
